+++ Realm/RLMMigration.mm
@@ -103,9 +103,12 @@
             // FIXME: replace with count of distinct once we support indexing
 
             // FIXME: support other types
-            tightdb::TableRef &table = objectSchema->_table;
+            tightdb::Table *table = objectSchema.table;
             NSUInteger count = table->size();
             if (primaryProperty.type == RLMPropertyTypeString) {
+                if (!table->has_search_index(primaryProperty.column)) {
+                    table->add_search_index(primaryProperty.column);
+                }
                 if (table->get_distinct_view(primaryProperty.column).size() != count) {
                     NSString *reason = [NSString stringWithFormat:@"Primary key property '%@' has duplicate values after migration.", primaryProperty.name];
                     @throw [NSException exceptionWithName:@"RLMException" reason:reason userInfo:nil];