+++ Realm/RLMProperty.mm
@@ -24,7 +24,9 @@
 #import "RLMSwiftSupport.h"
 #import "RLMUtil.hpp"
 
-@implementation RLMProperty
+@implementation RLMProperty {
+    NSString *_objcRawType;
+}
 
 - (instancetype)initWithName:(NSString *)name
                         type:(RLMPropertyType)type
@@ -37,13 +39,13 @@
         _objectClassName = objectClassName;
         _attributes = attributes;
         [self setObjcCodeFromType];
-        [self updateAccessorNames];
+        [self updateAccessors];
     }
 
     return self;
 }
 
--(void)updateAccessorNames {
+-(void)updateAccessors {
     // populate getter/setter names if generic
     if (!_getterName) {
         _getterName = _name;
@@ -136,13 +138,13 @@
                 Class cls = [RLMSchema classForString:_objectClassName];
                 if (!RLMIsSubclass(cls, RLMObject.class)) {
                     @throw [NSException exceptionWithName:@"RLMException"
-                                                   reason:[NSString stringWithFormat:@"'%@' is not supported as an RLMArray object type. The protocol for an RLMArray property must be defined with RLM_ARRAY_TYPE. See http://realm.io/docs/cocoa/latest/api/Protocols/RLM_ARRAY_TYPE.html for more information.", self.objectClassName]
+                                                   reason:[NSString stringWithFormat:@"'%@' is not supported as an RLMArray object type. RLMArrays can only contain instances of RLMObject subclasses. See http://realm.io/docs/cocoa/#to-many for more information.", self.objectClassName]
                                                  userInfo:nil];
                 }
             }
             else if (strcmp(code, "@\"NSNumber\"") == 0) {
                 @throw [NSException exceptionWithName:@"RLMException"
-                                               reason:[NSString stringWithFormat:@"'NSNumber' is not supported as an RLMObject property. Supported number types include int, long, float, double, and other primitive number types. See http://realm.io/docs/cocoa/latest/api/Constants/RLMPropertyType.html for all supported types."]
+                                               reason:[NSString stringWithFormat:@"'NSNumber' is not supported as an RLMObject property. Supported number types include int, long, float, double, and other primitive number types. See http://realm.io/docs/cocoa/api/Constants/RLMPropertyType.html for all supported types."]
                                              userInfo:nil];
             }
             else if (strcmp(code, "@\"RLMArray\"") == 0) {
@@ -158,7 +160,7 @@
                 Class cls = [RLMSchema classForString:className];
                 if (!RLMIsSubclass(cls, RLMObject.class)) {
                     @throw [NSException exceptionWithName:@"RLMException"
-                                                   reason:[NSString stringWithFormat:@"'%@' is not supported as an RLMObject property. All properties must be primitives, NSString, NSDate, NSData, RLMArray, or subclasses of RLMObject. See http://realm.io/docs/cocoa/latest/api/Classes/RLMObject.html for more information.", self.objectClassName]
+                                                   reason:[NSString stringWithFormat:@"'%@' is not supported as an RLMObject property. All properties must be primitives, NSString, NSDate, NSData, RLMArray, or subclasses of RLMObject. See http://realm.io/docs/cocoa/api/Classes/RLMObject.html for more information.", self.objectClassName]
                                                  userInfo:nil];
                 }
 
@@ -241,7 +243,7 @@
     }
 
     // update getter/setter names
-    [self updateAccessorNames];
+    [self updateAccessors];
 
     return self;
 }
@@ -269,11 +271,25 @@
     }
 
     // update getter/setter names
-    [self updateAccessorNames];
+    [self updateAccessors];
 
     return self;
 }
 
+- (id)copyWithZone:(NSZone *)zone {
+    RLMProperty *prop = [[RLMProperty allocWithZone:zone] init];
+    prop->_name = _name;
+    prop->_type = _type;
+    prop->_objcType = _objcType;
+    prop->_objectClassName = _objectClassName;
+    prop->_attributes = _attributes;
+    prop->_getterName = _getterName;
+    prop->_setterName = _setterName;
+    prop->_isPrimary = _isPrimary;
+    [prop updateAccessors];
+    
+    return prop;
+}
 
 -(BOOL)isEqualToProperty:(RLMProperty *)prop {
     return [_name isEqualToString:prop.name] && _type == prop.type && prop.isPrimary == _isPrimary &&