+++ Realm/RLMSchema.h
@@ -58,4 +58,9 @@
  */
 - (RLMObjectSchema *)objectForKeyedSubscript:(id <NSCopying>)className;
 
+/**
+ Returns YES if schema are equal
+ */
+- (BOOL)isEqualToSchema:(RLMSchema *)schema;
+
 @end