+++ Realm/Tests/MigrationTests.mm
@@ -43,6 +43,16 @@ extern "C" {
 }
 @end
 
+@interface MigrationStringPrimaryKeyObject : RLMObject
+@property NSString * stringCol;
+@end
+
+@implementation MigrationStringPrimaryKeyObject
++ (NSString *)primaryKey {
+    return @"stringCol";
+}
+@end
+
 @interface MigrationTests : RLMTestCase
 @end
 
@@ -51,10 +61,20 @@ extern "C" {
 - (RLMRealm *)realmWithSingleObject:(RLMObjectSchema *)objectSchema {
     // modify object schema to use RLMObject class (or else bad accessors will get created)
     objectSchema.objectClass = RLMObject.class;
+    objectSchema.accessorClass = RLMObject.class;
 
     RLMSchema *schema = [[RLMSchema alloc] init];
     schema.objectSchema = @[objectSchema];
     return [self dynamicRealmWithTestPathAndSchema:schema];
+}
+
+- (void)testSchemaVersion {
+    [RLMRealm setSchemaVersion:1 withMigrationBlock:^(__unused RLMMigration *migration,
+                                                      __unused NSUInteger oldSchemaVersion) {
+    }];
+
+    RLMRealm *defaultRealm = [RLMRealm defaultRealm];
+    XCTAssertEqual(1U, RLMRealmSchemaVersion(defaultRealm));
 }
 
 - (void)testAddingProperty {
@@ -92,6 +112,9 @@ extern "C" {
     MigrationObject *mig1 = [MigrationObject allObjectsInRealm:realm][1];
     XCTAssertEqual(mig1.intCol, 2, @"Int column should have value 2");
     XCTAssertEqualObjects(mig1.stringCol, @"2", @"String column should be populated");
+
+    [RLMRealm setSchemaVersion:0 withMigrationBlock:nil];
+    XCTAssertThrows([RLMRealm migrateRealmAtPath:RLMTestRealmPath()]);
 }
 
 
@@ -223,6 +246,51 @@ extern "C" {
     [RLMRealm migrateRealmAtPath:RLMTestRealmPath()];
 }
 
+- (void)testStringPrimaryKeyMigration {
+    // make string an int
+    RLMObjectSchema *objectSchema = [RLMObjectSchema schemaForObjectClass:MigrationStringPrimaryKeyObject.class];
+    objectSchema.primaryKeyProperty.isPrimary = NO;
+    objectSchema.primaryKeyProperty = nil;
+
+    // create realm with old schema and populate
+    RLMRealm *realm = [self realmWithSingleObject:objectSchema];
+    [realm beginWriteTransaction];
+    [realm createObject:MigrationStringPrimaryKeyObject.className withObject:@[@"1"]];
+    [realm createObject:MigrationStringPrimaryKeyObject.className withObject:@[@"2"]];
+    [realm commitWriteTransaction];
+
+    // apply migration
+    [RLMRealm setSchemaVersion:1 withMigrationBlock:^(__unused RLMMigration *migration, __unused NSUInteger oldSchemaVersion) {
+        [migration enumerateObjects:@"MigrationStringPrimaryKeyObject" block:^(__unused RLMObject *oldObject, RLMObject *newObject) {
+            newObject[@"stringCol"] = [[NSUUID UUID] UUIDString];
+        }];
+    }];
+    [RLMRealm migrateRealmAtPath:RLMTestRealmPath()];
+}
+
+- (void)testStringPrimaryKeyNoIndexMigration {
+    // make string an int
+    RLMObjectSchema *objectSchema = [RLMObjectSchema schemaForObjectClass:MigrationStringPrimaryKeyObject.class];
+
+    // create without search index
+    objectSchema.primaryKeyProperty.attributes = 0;
+
+    // create realm with old schema and populate
+    RLMRealm *realm = [self realmWithSingleObject:objectSchema];
+    [realm beginWriteTransaction];
+    [realm createObject:MigrationStringPrimaryKeyObject.className withObject:@[@"1"]];
+    [realm createObject:MigrationStringPrimaryKeyObject.className withObject:@[@"2"]];
+    [realm commitWriteTransaction];
+
+    // apply migration
+    [RLMRealm setSchemaVersion:1 withMigrationBlock:^(__unused RLMMigration *migration, __unused NSUInteger oldSchemaVersion) {
+        [migration enumerateObjects:@"MigrationStringPrimaryKeyObject" block:^(__unused RLMObject *oldObject, RLMObject *newObject) {
+            newObject[@"stringCol"] = [[NSUUID UUID] UUIDString];
+        }];
+    }];
+    [RLMRealm migrateRealmAtPath:RLMTestRealmPath()];
+}
+
 - (void)testDuplicatePrimaryKeyMigration {
     // make string an int
     RLMObjectSchema *objectSchema = [RLMObjectSchema schemaForObjectClass:MigrationPrimaryKeyObject.class];
@@ -331,8 +399,13 @@ extern "C" {
 }
 
 - (void)testRearrangeProperties {
-    // create realm with the properties reversed
     @autoreleasepool {
+        // create object in default realm
+        [[RLMRealm defaultRealm] transactionWithBlock:^{
+            [CircleObject createInDefaultRealmWithObject:@[@"data", NSNull.null]];
+        }];
+
+        // create realm with the properties reversed
         RLMSchema *schema = [[RLMSchema sharedSchema] copy];
         RLMObjectSchema *objectSchema = schema[@"CircleObject"];
         objectSchema.properties = @[objectSchema.properties[1], objectSchema.properties[0]];
@@ -353,6 +426,23 @@ extern "C" {
     XCTAssertNoThrow(obj.data = @"new data");
     XCTAssertNoThrow(obj.next = obj);
     [realm commitWriteTransaction];
+
+    // open the default Realm and make sure accessors with alternate ordering work
+    CircleObject *defaultObj = [[CircleObject allObjects] firstObject];
+    XCTAssertEqualObjects(defaultObj.data, @"data");
+
+    // test object from other realm still works
+    XCTAssertEqualObjects(obj.data, @"new data");
+
+    // verify schema for both objects
+    NSArray *properties = defaultObj.objectSchema.properties;
+    for (NSUInteger i = 0; i < properties.count; i++) {
+        XCTAssertEqual([properties[i] column], i);
+    }
+    properties = obj.objectSchema.properties;
+    for (NSUInteger i = 0; i < properties.count; i++) {
+        XCTAssertEqual([properties[i] column], i);
+    }
 }
 
 @end