+++ Realm/Tests/ObjectTests.m
@@ -1004,15 +1004,15 @@ RLM_ARRAY_TYPE(PrimaryIntObject);
 
 - (void)testIsDeleted {
     StringObject *obj1 = [[StringObject alloc] initWithObject:@[@"a"]];
-    XCTAssertEqual(obj1.deletedFromRealm, NO);
+    XCTAssertEqual(obj1.invalidated, NO);
 
     RLMRealm *realm = [self realmWithTestPath];
     [realm beginWriteTransaction];
     [realm addObject:obj1];
     StringObject *obj2 = [StringObject createInRealm:realm withObject:@[@"b"]];
 
-    XCTAssertEqual([obj1 isDeletedFromRealm], NO);
-    XCTAssertEqual(obj2.deletedFromRealm, NO);
+    XCTAssertEqual([obj1 isInvalidated], NO);
+    XCTAssertEqual(obj2.invalidated, NO);
 
     [realm commitWriteTransaction];
 
@@ -1021,14 +1021,14 @@ RLM_ARRAY_TYPE(PrimaryIntObject);
     [realm deleteObject:obj1];
     [realm deleteObject:obj2];
 
-    XCTAssertEqual(obj1.deletedFromRealm, YES);
-    XCTAssertEqual(obj2.deletedFromRealm, YES);
+    XCTAssertEqual(obj1.invalidated, YES);
+    XCTAssertEqual(obj2.invalidated, YES);
 
     XCTAssertThrows([realm addObject:obj1], @"Adding deleted object should throw");
     
     [realm commitWriteTransaction];
 
-    XCTAssertEqual(obj1.deletedFromRealm, YES);
+    XCTAssertEqual(obj1.invalidated, YES);
     XCTAssertNil(obj1.realm, @"Realm should be nil after deletion");
 }