+++ Realm/Tests/RLMTestCase.m
@@ -110,7 +110,7 @@ static void RLMDeleteRealmFilesAtPath(NSString *path) {
 }
 
 - (RLMRealm *)realmWithTestPathAndSchema:(RLMSchema *)schema {
-    return [RLMRealm realmWithPath:RLMTestRealmPath() readOnly:NO inMemory:NO dynamic:NO schema:schema error:nil];
+    return [RLMRealm realmWithPath:RLMTestRealmPath() readOnly:NO inMemory:NO dynamic:YES schema:schema error:nil];
 }
 
 - (RLMRealm *)dynamicRealmWithTestPathAndSchema:(RLMSchema *)schema {