+++ tools/RealmBrowser/RealmBrowser/Classes/RLMRealmBrowserWindowController.m
@@ -55,7 +55,7 @@ NSString * const kRealmKeyOutlineWidthForRealm = @"OutlineWidthForRealm:%@";
 - (void)windowDidLoad
 {
     navigationStack = [[RLMNavigationStack alloc] init];
-    self.window.alphaValue = 0.0;
+    [self realmDidLoad];
 }
 
 #pragma mark - RLMViewController Overrides
@@ -78,7 +78,6 @@ NSString * const kRealmKeyOutlineWidthForRealm = @"OutlineWidthForRealm:%@";
     [self.splitView setAutosaveName:[NSString stringWithFormat:kRealmKeyOutlineWidthForRealm, realmPath]];
     
     [self reloadAfterEdit];
-    self.window.alphaValue = 1.0;
 }
 
 #pragma mark - Public methods - Accessors